requirements.txt
# Core dependencies
python-dotenv>=1.0.0
PyMuPDF>=1.23.0
pdfplumber>=0.10.0

# Embeddings and Vector Store
faiss-cpu>=1.7.4
google-generativeai>=0.8.0
sentence-transformers>=2.2.0

# LangChain/LangGraph ecosystem
langchain>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
langchain-community>=0.3.0

# Hybrid search (BM25)
rank-bm25>=0.2.2

# Reranking
sentence-transformers>=2.2.0

# UI
streamlit>=1.29.0

# Export functionality
python-docx>=1.1.0
reportlab>=4.0.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
